Change the gerrit data persisting workflow
The gerrit server was creating binary files in the docker container,
that were supposed to be mounted locally and checked into the repo.
This was a bad idea, amongst other things, because this makes it harder
to track whether unintended data is uploaded in a commit and also
no 'diff' can be created for binary data blobs in git, which means that
the entire file gets re-uploaded whenever something is commited, which
might bloat the repo.

The NEW APPROACH is to create rakefiles that will get run by the main
rakefile in the root directory, and it will create all the data once
the server is started. Just run 'bundle exec rake prepare_gerrit_data'
before running the tests.